feat: refactor and enable keplr
