+++ projects/portal/src/app/models/keplr/keplr.service.ts
@@ -1,125 +1,33 @@
-import { ConfigService } from '../config.service';
-import { CosmosSDKService } from '../cosmos-sdk.service';
+import { StoredWallet } from '../wallets/wallet.model';
+import { KeplrInfrastructureService } from './keplr.infrastructure.service';
 import { Injectable } from '@angular/core';
-import { ChainInfo, Key, Window as KeplrWindow } from '@keplr-wallet/types';
+import { cosmosclient, proto } from '@cosmos-client/core';
 
-export interface signKeplr {
-  authInfoBytes: Uint8Array;
-  bodyBytes: Uint8Array;
-  signature: Uint8Array;
+export interface IKeplrInfrastructureService {
+  connectWallet: () => Promise<StoredWallet | null | undefined>;
+  signTx: (
+    txBuilder: cosmosclient.TxBuilder,
+    signerBaseAccount: proto.cosmos.auth.v1beta1.BaseAccount,
+  ) => Promise<cosmosclient.TxBuilder>;
 }
-
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-empty-interface
-  interface Window extends KeplrWindow {}
-}
-
 @Injectable({
   providedIn: 'root',
 })
 export class KeplrService {
-  constructor(private readonly cosmosSDK: CosmosSDKService, private configService: ConfigService) {}
+  private readonly iKeplrInfrastructureService: IKeplrInfrastructureService;
 
-  async getKey(): Promise<Key | undefined> {
-    if (!window.keplr) {
-      alert('Please install keplr extension');
-      return;
-    } else {
-      const chainID = this.configService.configs[0].chainID;
-      await window.keplr?.enable(chainID);
-
-      const key = await window.keplr?.getKey(chainID);
-      return key;
-    }
+  constructor(readonly keplrInfrastructureService: KeplrInfrastructureService) {
+    this.iKeplrInfrastructureService = this.keplrInfrastructureService;
   }
 
-  async suggestChain(): Promise<void> {
-    if (!window.keplr) {
-      alert('Please install keplr extension');
-      return;
-    } else {
-      const chainId = this.configService.configs[0].chainID;
-      const chainName = this.configService.configs[0].chainName;
-      const rpc = this.configService.configs[0].websocketURL;
-      const rest = this.configService.configs[0].restURL;
-      const bip44 = { coinType: 118 };
-      const bech32Config = {
-        bech32PrefixAccAddr: this.configService.configs[0].bech32Prefix?.accAddr!,
-        bech32PrefixAccPub: this.configService.configs[0].bech32Prefix?.accPub!,
-        bech32PrefixValAddr: this.configService.configs[0].bech32Prefix?.valAddr!,
-        bech32PrefixValPub: this.configService.configs[0].bech32Prefix?.valPub!,
-        bech32PrefixConsAddr: this.configService.configs[0].bech32Prefix?.consAddr!,
-        bech32PrefixConsPub: this.configService.configs[0].bech32Prefix?.consPub!,
-      };
-      const currencies = [
-        {
-          coinDenom: 'GUU',
-          coinMinimalDenom: 'uguu',
-          coinDecimals: 6,
-          coinGeckoId: 'ununifi',
-        },
-      ];
-      const feeCurrencies = [
-        {
-          coinDenom: 'GUU',
-          coinMinimalDenom: 'uguu',
-          coinDecimals: 6,
-          coinGeckoId: 'ununifi',
-        },
-      ];
-      const stakeCurrency = {
-        coinDenom: 'GUU',
-        coinMinimalDenom: 'uguu',
-        coinDecimals: 6,
-        coinGeckoId: 'ununifi',
-      };
-      const coinType = 118;
-      const gasPriceStep = {
-        low: 0,
-        average: 0.01,
-        high: 0.03,
-      };
-      const chainInfo: ChainInfo = {
-        chainId,
-        chainName,
-        rpc,
-        rest,
-        bip44,
-        bech32Config,
-        currencies,
-        feeCurrencies,
-        stakeCurrency,
-        coinType,
-        gasPriceStep,
-      };
-      await window.keplr?.experimentalSuggestChain(chainInfo);
-    }
+  async connectWallet(): Promise<StoredWallet | null | undefined> {
+    return await this.iKeplrInfrastructureService.connectWallet();
   }
 
-  async signDirect(
-    signer: string,
-    bodyBytes: Uint8Array,
-    authInfoBytes: Uint8Array,
-    accountNumber: Long,
-  ): Promise<signKeplr | undefined> {
-    if (!window.keplr) {
-      alert('Please install keplr extension');
-      return;
-    } else {
-      const chainId = this.configService.configs[0].chainID;
-      await window.keplr?.enable(chainId);
-      const directSignResponse = await window.keplr.signDirect(chainId, signer, {
-        bodyBytes,
-        authInfoBytes,
-        chainId,
-        accountNumber,
-      });
-      const signKeplr: signKeplr = {
-        authInfoBytes: directSignResponse.signed.authInfoBytes,
-        bodyBytes: directSignResponse.signed.bodyBytes,
-        signature: Uint8Array.from(Buffer.from(directSignResponse.signature.signature, 'base64')),
-      };
-      return signKeplr;
-    }
+  async signTx(
+    txBuilder: cosmosclient.TxBuilder,
+    signerBaseAccount: proto.cosmos.auth.v1beta1.BaseAccount,
+  ): Promise<cosmosclient.TxBuilder> {
+    return await this.iKeplrInfrastructureService.signTx(txBuilder, signerBaseAccount);
   }
 }
