--- projects/portal/src/app/views/dialogs/wallets/keplr/keplr-import-wallet-dialog/keplr-import-wallet-dialog.component.html
@@ -1,29 +1,0 @@
-<mat-dialog-content>
-  <div class="flex flex-col items-center overflow-y-auto">
-    <img class="m-6 w-12 h-12" src="assets/favicon.png" alt="UnUniFi logo" />
-    <div class="font-bold text-xl m-3">Connect Keplr Wallet</div>
-    <div class="text-base m-3">Start earning high yield on your crypto</div>
-    <ng-container *ngIf="storedWallet$ | async as storedWallet">
-      <ng-container *ngIf="storedWallet">
-        <mat-nav-list>
-          <mat-list-item (click)="onClickWallet(storedWallet)">
-            <mat-icon matListIcon [ngStyle]="{ color: getColorCode(storedWallet) }"
-              >circle</mat-icon
-            >
-            <div mat-line>
-              <span>{{ storedWallet.id }}</span>
-              <ng-container *ngIf="currentStoredWallet$ | async as currentStoredWallet">
-                <ng-container *ngIf="currentStoredWallet.id === storedWallet.id">
-                  <mat-icon color="primary">check</mat-icon>
-                </ng-container>
-              </ng-container>
-            </div>
-            <div mat-line>
-              <span>{{ storedWallet.address }}</span>
-            </div>
-          </mat-list-item>
-        </mat-nav-list>
-      </ng-container>
-    </ng-container>
-  </div>
-</mat-dialog-content>
