--- projects/portal/src/app/views/dialogs/wallets/keplr/keplr-import-wallet-dialog/keplr-import-wallet-dialog.module.ts
@@ -1,11 +1,0 @@
-import { MaterialModule } from '../../../../material.module';
-import { KeplrImportWalletDialogComponent } from './keplr-import-wallet-dialog.component';
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-
-@NgModule({
-  declarations: [KeplrImportWalletDialogComponent],
-  imports: [CommonModule, MaterialModule],
-  exports: [KeplrImportWalletDialogComponent],
-})
-export class KeplrImportWalletDialogModule {}
